requirements.txt
fastapi
uvicorn[standard]
pydantic
scikit-learn
numpy
pandas
kagglehub
python-dotenv
